LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/Create-invalid-urls.any-expected.txt


PASS Url is /echo - should throw SYNTAX_ERR
PASS Url is a mail address - should throw SYNTAX_ERR
PASS Url is about:blank - should throw SYNTAX_ERR
PASS Url is ?test - should throw SYNTAX_ERR
PASS Url is #test - should throw SYNTAX_ERR
